fix: change base structure
