+++ src/bin/statique.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
 DIR="$( cd "$(dirname "$0")" ; pwd -P )"
-java -cp "$DIR/../lib/*" ch.heigvd.statique.Statique "$@"
+java -cp "$DIR/../lib/*" ch.heig.statique.Statique "$@"
