+++ src/main/java/ch/heig/statique/Commands/Clean.java
@@ -1,16 +1,16 @@
-package ch.heig.dil.Commands;
+package ch.heig.statique.Commands;
 
 import picocli.CommandLine;
 
 import java.util.concurrent.Callable;
 
 @CommandLine.Command(
-        name = "serve",
+        name = "clean",
         mixinStandardHelpOptions = true,
         version = "0.1",
-        description = "Serve a static site"
+        description = "Clean a static site"
 )
-public class Serve implements Callable<Integer> {
+public class Clean implements Callable<Integer> {
 
     @Override
     public Integer call() throws Exception {
